requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-appconf==1.0.4
django-compressor==2.4.1
gunicorn==20.1.0
python-dotenv==0.19.0
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
